test/main: refactor process-wide setup/shutdown. (#6858)
We have quite a few third party libraries that need to have process wide
setup/shutdown (including nghttp2, c-ares, gRPC, libevent). Ideally we
have a single source of truth for this sequence.

This PR is motivated by a TSAN race in CI reported in #6839, where the
DnsResolverImpl was doing its own c-ares setup/shutdown, racing with
gRPC library.

While we're mucking around in this code, also fix #6689.

Risk level: Low
Testing: bazel test --config=clang-tsan -c dbg
  //test/server:server_fuzz_test --runs_per_test=5000

Fixes #6839
Fixes #6689
