+++ source/exe/process_wide.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <cstdint>
+
+namespace Envoy {
+
+// Process-wide lifecycle events for global state in third-party dependencies,
+// e.g. gRPC, c-ares. There should only ever be a single instance of this.
+class ProcessWide {
+public:
+  ProcessWide();
+  ~ProcessWide();
+
+private:
+  uint32_t initialization_depth_;
+};
+
+} // namespace Envoy
