+++ source/extensions/stat_sinks/metrics_service/config.cc
@@ -19,7 +19,6 @@ namespace MetricsService {
 
 Stats::SinkPtr MetricsServiceSinkFactory::createStatsSink(const Protobuf::Message& config,
                                                           Server::Instance& server) {
-  RELEASE_ASSERT(validateProtoDescriptors(), "");
   const auto& sink_config =
       MessageUtil::downcastAndValidate<const envoy::config::metrics::v2::MetricsServiceConfig&>(
           config);
