+++ test/fuzz/BUILD
@@ -39,6 +39,7 @@ envoy_cc_test_library(
         "//source/common/common:utility_lib",
         "//source/common/event:libevent_lib",
         "//source/common/http/http2:codec_lib",
+        "//source/exe:process_wide_lib",
         "//test/test_common:environment_lib",
     ],
 )
