+++ test/server/config_validation/dispatcher_test.cc
@@ -22,8 +22,6 @@ namespace Envoy {
 class ConfigValidation : public testing::TestWithParam<Network::Address::IpVersion> {
 public:
   ConfigValidation() {
-    Event::Libevent::Global::initialize();
-
     validation_ = std::make_unique<Api::ValidationImpl>(Thread::threadFactoryForTest(),
                                                         stats_store_, test_time_.timeSystem(),
                                                         Filesystem::fileSystemForTest());
