+++ test/server/server_fuzz_test.cc
@@ -68,8 +68,6 @@ DEFINE_PROTO_FUZZER(const envoy::config::bootstrap::v2::Bootstrap& input) {
   DangerousDeprecatedTestTime test_time;
   Fuzz::PerTestEnvironment test_env;
 
-  RELEASE_ASSERT(validateProtoDescriptors(), "");
-
   {
     const std::string bootstrap_path = test_env.temporaryPath("bootstrap.pb_text");
     std::ofstream bootstrap_file(bootstrap_path);
